Second Alpha 0.2

+++ Sci-Fi-Skirmish-main/Sci-Fi Skirmish/1/JansParadise/src/main/java/jansparadise/jansparadise/models/PlayerStats.java
@@ -88,7 +88,7 @@ public class PlayerStats {
     }
 
     public long getUwu() {
-        return (long) uwu;
+        return uwu;
     }
 
     public int getCommon_crates() {
